Updating user management to use new forms
